llu for printing ticks
git-svn-id: https://trigonak-thesis.googlecode.com/svn/exjobb/iRCCEdeddsl@4028 dd814760-9b98-ad33-1d6a-b699670a4318

+++ include/measurements.h
@@ -218,11 +218,11 @@ while (0);
                 int msa = (int) trunc((secsa - sa) * 1000);
                 int usa = (int) trunc(((secsa - sa) * 1000000) - (msa * 1000));
                 int nsa = (int) trunc(((secsa - sa) * 1000000000) - (msa * 1000000) - (usa * 1000));
-                printf("  [%2.1f%%] \tsamples: %-16llu | time: %3d %3d %3d %3d | avg time: %3d %3d %3d %3d \t(s ms us ns) || ticks: %d\n",
+                printf("  [%2.1f%%] \tsamples: %-16llu | time: %3d %3d %3d %3d | avg time: %3d %3d %3d %3d \t(s ms us ns) || ticks: %llu\n",
                         ticks_perc, total_samples[i],
                         s, ms, us, ns,
                         sa, msa, usa, nsa,
-                        (long long int) total_sum_ticks[i]);
+                        total_sum_ticks[i]);
             }
         }
